refactor: streamline report generation and enhance error handling
- Updated the report generation process to generate a comprehensive report once and reuse it for printing and saving, improving efficiency.
- Refactored the `ConsoleReporter` and `ReporterService` to utilize shared functions for computing problem statistics and generating report summaries, reducing code duplication.
- Enhanced error reporting by standardizing the handling of parsing failures and type resolution warnings across the codebase.
