+++ packages/reporter/src/reporter-service.ts
@@ -8,9 +8,8 @@ import { resolve } from "node:path";
 import { blue, green, red, yellow } from "colorette";
 import { PACKAGE_VERSION } from "./constants.ts";
 import type { ReporterBase } from "./reporter-base.ts";
-import type { GenerationReport, ProblemEntry, ReportStatistics } from "./types/index.ts";
-import { ProblemCategory, ProblemSeverity } from "./types/index.ts";
-import { determineGenerationStatus } from "./types/report.ts";
+import type { GenerationReport, ProblemEntry } from "./types/index.ts";
+import { computeProblemStatistics, generateReportSummary, groupProblemsByCategory } from "./types/report.ts";
 
 /**
  * Centralized service for managing multiple Reporter instances
@@ -84,8 +83,7 @@ export class ReporterService {
 		const allProblems: ProblemEntry[] = [];
 
 		for (const reporter of this.reporters.values()) {
-			const report = reporter.generateReport();
-			allProblems.push(...report.problems);
+			allProblems.push(...reporter.getProblems());
 		}
 
 		return allProblems;
@@ -94,235 +92,14 @@ export class ReporterService {
 	/**
 	 * Generate comprehensive statistics from all reporters
 	 */
-	private generateComprehensiveStatistics(): ReportStatistics {
-		const allProblems = this.collectAllProblems();
-
-		// If we have no reporters, return empty stats
-		if (this.reporters.size === 0) {
-			return {
-				bySeverity: {
-					[ProblemSeverity.DEBUG]: 0,
-					[ProblemSeverity.INFO]: 0,
-					[ProblemSeverity.WARNING]: 0,
-					[ProblemSeverity.ERROR]: 0,
-					[ProblemSeverity.CRITICAL]: 0,
-				},
-				byCategory: {
-					[ProblemCategory.TYPE_RESOLUTION]: 0,
-					[ProblemCategory.PARSING_FAILURE]: 0,
-					[ProblemCategory.GENERATION_FAILURE]: 0,
-					[ProblemCategory.TYPE_CONFLICT]: 0,
-					[ProblemCategory.DEPENDENCY_ISSUE]: 0,
-					[ProblemCategory.CONFIGURATION]: 0,
-					[ProblemCategory.IO_ERROR]: 0,
-					[ProblemCategory.GENERAL]: 0,
-				},
-				byModule: {},
-				totalProblems: 0,
-				mostProblematicModules: [],
-				typeStatistics: {
-					commonUnresolvedTypes: [],
-					commonTypeConflicts: [],
-					problematicNamespaces: [],
-				},
-				startTime: new Date(),
-				endTime: new Date(),
-				durationMs: 0,
-			};
-		}
-
-		// Get the first reporter's report as base for timing
-		const firstReport = this.reporters.values().next().value?.generateReport();
-		const startTime = firstReport?.statistics.startTime || new Date();
-
-		// Use current time as end time
-		const endTime = new Date();
-		const durationMs = endTime.getTime() - startTime.getTime();
-
-		// Aggregate statistics
-		const bySeverity: Record<ProblemSeverity, number> = {
-			[ProblemSeverity.DEBUG]: 0,
-			[ProblemSeverity.INFO]: 0,
-			[ProblemSeverity.WARNING]: 0,
-			[ProblemSeverity.ERROR]: 0,
-			[ProblemSeverity.CRITICAL]: 0,
-		};
-
-		const byCategory: Record<ProblemCategory, number> = {
-			[ProblemCategory.TYPE_RESOLUTION]: 0,
-			[ProblemCategory.PARSING_FAILURE]: 0,
-			[ProblemCategory.GENERATION_FAILURE]: 0,
-			[ProblemCategory.TYPE_CONFLICT]: 0,
-			[ProblemCategory.DEPENDENCY_ISSUE]: 0,
-			[ProblemCategory.CONFIGURATION]: 0,
-			[ProblemCategory.IO_ERROR]: 0,
-			[ProblemCategory.GENERAL]: 0,
-		};
-
-		const byModule: Record<string, number> = {};
-
-		// Type-specific tracking
-		const unresolvedTypes: Record<string, { count: number; namespaces: Set<string>; sourceModules: Set<string> }> = {};
-		const typeConflicts: Record<string, { count: number; examples: Set<string> }> = {};
-		const namespaceProblems: Record<string, { count: number; types: Set<string> }> = {};
-
-		for (const problem of allProblems) {
-			bySeverity[problem.severity] = (bySeverity[problem.severity] || 0) + 1;
-			byCategory[problem.category] = (byCategory[problem.category] || 0) + 1;
-			byModule[problem.module] = (byModule[problem.module] || 0) + 1;
-
-			// Track type resolution problems
-			if (problem.category === ProblemCategory.TYPE_RESOLUTION && problem.typeName) {
-				if (!unresolvedTypes[problem.typeName]) {
-					unresolvedTypes[problem.typeName] = { count: 0, namespaces: new Set(), sourceModules: new Set() };
-				}
-				unresolvedTypes[problem.typeName].count++;
-				if (problem.location) {
-					unresolvedTypes[problem.typeName].namespaces.add(problem.location);
-				}
-				const sourceModule = problem.metadata?.sourceModule as string | undefined;
-				if (sourceModule) {
-					unresolvedTypes[problem.typeName].sourceModules.add(sourceModule);
-				}
-
-				// Track namespace problems
-				if (problem.location) {
-					if (!namespaceProblems[problem.location]) {
-						namespaceProblems[problem.location] = { count: 0, types: new Set() };
-					}
-					namespaceProblems[problem.location].count++;
-					namespaceProblems[problem.location].types.add(problem.typeName);
-				}
-			}
-
-			// Track type conflicts
-			if (problem.category === ProblemCategory.TYPE_CONFLICT && problem.metadata?.conflictType) {
-				const conflictType = problem.metadata.conflictType as string;
-				if (!typeConflicts[conflictType]) {
-					typeConflicts[conflictType] = { count: 0, examples: new Set() };
-				}
-				typeConflicts[conflictType].count++;
-				if (problem.typeName) {
-					typeConflicts[conflictType].examples.add(problem.typeName);
-				}
-			}
-		}
-
-		// Convert to arrays and sort
-		const commonUnresolvedTypes = Object.entries(unresolvedTypes)
-			.map(([type, data]) => ({
-				type,
-				count: data.count,
-				namespaces: Array.from(data.namespaces),
-				sourceModules: Array.from(data.sourceModules),
-			}))
-			.sort((a, b) => b.count - a.count)
-			.slice(0, 20);
-
-		const commonTypeConflicts = Object.entries(typeConflicts)
-			.map(([conflictType, data]) => ({
-				conflictType,
-				count: data.count,
-				examples: Array.from(data.examples).slice(0, 5),
-			}))
-			.sort((a, b) => b.count - a.count);
-
-		const problematicNamespaces = Object.entries(namespaceProblems)
-			.map(([namespace, data]) => ({
-				namespace,
-				problems: data.count,
-				types: Array.from(data.types).slice(0, 10),
-			}))
-			.sort((a, b) => b.problems - a.problems)
-			.slice(0, 10);
-
-		const mostProblematicModules = Object.entries(byModule)
-			.sort(([, a], [, b]) => b - a)
-			.slice(0, 10)
-			.map(([module, count]) => ({ module, count }));
-
-		return {
-			bySeverity,
-			byCategory,
-			byModule,
-			totalProblems: allProblems.length,
-			mostProblematicModules,
-			typeStatistics: {
-				commonUnresolvedTypes,
-				commonTypeConflicts,
-				problematicNamespaces,
-			},
-			startTime,
-			endTime,
-			durationMs,
-		};
-	}
-
 	/**
 	 * Generate comprehensive report from all reporters
 	 */
 	public generateComprehensiveReport(): GenerationReport {
-		const statistics = this.generateComprehensiveStatistics();
 		const allProblems = this.collectAllProblems();
-
-		// Generate problems by category
-		const problemsByCategory = Object.values(ProblemCategory).reduce(
-			(acc, category) => {
-				acc[category] = [];
-				return acc;
-			},
-			{} as Record<ProblemCategory, ProblemEntry[]>,
-		);
-
-		for (const problem of allProblems) {
-			problemsByCategory[problem.category].push(problem);
-		}
-
-		// Generate summary
-		const status = determineGenerationStatus(statistics.bySeverity, statistics.byCategory);
-
-		const keyIssues: string[] = [];
-		const recommendations: string[] = [];
-
-		// Analyze key issues across all modules
-		const typeResolutionCount = statistics.byCategory[ProblemCategory.TYPE_RESOLUTION] || 0;
-		const parsingFailureCount = statistics.byCategory[ProblemCategory.PARSING_FAILURE] || 0;
-		const generationFailureCount = statistics.byCategory[ProblemCategory.GENERATION_FAILURE] || 0;
-		const conflictCount = statistics.byCategory[ProblemCategory.TYPE_CONFLICT] || 0;
-
-		if (typeResolutionCount > 0) {
-			keyIssues.push(
-				`${typeResolutionCount} type resolution warnings across all modules (produce 'never' type in output)`,
-			);
-			recommendations.push(
-				"Unresolved types produce 'never' in output — these are typically non-introspectable types or missing GIR dependencies",
-			);
-		}
-
-		if (parsingFailureCount > 0) {
-			keyIssues.push(`${parsingFailureCount} parsing failures encountered`);
-			recommendations.push("Check GIR file syntax and ensure proper introspection data");
-		}
-
-		if (generationFailureCount > 0) {
-			keyIssues.push(`${generationFailureCount} generation failures occurred`);
-			recommendations.push("Review template configuration and output settings");
-		}
-
-		if (conflictCount > 0) {
-			keyIssues.push(`${conflictCount} type conflicts detected`);
-			recommendations.push(
-				"Type conflicts are handled automatically — conflicting members are omitted or use union types",
-			);
-		}
-
-		if (keyIssues.length === 0 && statistics.totalProblems > 0) {
-			keyIssues.push(`${statistics.totalProblems} minor issues detected across all modules`);
-		}
-
-		if (recommendations.length === 0 && statistics.totalProblems > 0) {
-			recommendations.push("Review detailed problem list for specific improvement opportunities");
-		}
+		const startTime = this.reporters.values().next().value?.getProblems()[0]?.timestamp || new Date();
+		const statistics = computeProblemStatistics(allProblems, startTime);
+		const summary = generateReportSummary(statistics);
 
 		return {
 			metadata: {
@@ -331,24 +108,20 @@ export class ReporterService {
 			},
 			statistics,
 			problems: allProblems,
-			problemsByCategory,
-			summary: {
-				status,
-				keyIssues,
-				recommendations,
-			},
+			problemsByCategory: groupProblemsByCategory(allProblems),
+			summary,
 		};
 	}
 
 	/**
 	 * Save comprehensive report to file
 	 */
-	public async saveComprehensiveReport(outputPath?: string): Promise<void> {
+	public async saveComprehensiveReport(outputPath?: string, precomputedReport?: GenerationReport): Promise<void> {
 		if (!this.config.enabled) {
 			return;
 		}
 
-		const report = this.generateComprehensiveReport();
+		const report = precomputedReport || this.generateComprehensiveReport();
 		const filePath = outputPath || this.config.outputPath;
 
 		try {
@@ -363,8 +136,8 @@ export class ReporterService {
 	/**
 	 * Print comprehensive summary to console
 	 */
-	public printComprehensiveSummary(): void {
-		const report = this.generateComprehensiveReport();
+	public printComprehensiveSummary(precomputedReport?: GenerationReport): void {
+		const report = precomputedReport || this.generateComprehensiveReport();
 		const { statistics, summary } = report;
 
 		console.log(`\n${"=".repeat(60)}`);
